Fix test and make noop initializers singleton
