+++ gradle/scripts/dependencyDefinitions.gradle
@@ -117,8 +117,8 @@ ext.externalDependency = [
     "bytemanBmunit": "org.jboss.byteman:byteman-bmunit:" + bytemanVersion,
     "bcpgJdk15on": "org.bouncycastle:bcpg-jdk15on:1.52",
     "bcprovJdk15on": "org.bouncycastle:bcprov-jdk15on:1.52",
-    "calciteCore": "org.apache.calcite:calcite-core:1.6.0",
-    "calciteAvatica": "org.apache.calcite:calcite-avatica:1.6.0",
+    "calciteCore": "org.apache.calcite:calcite-core:1.3.0-incubating",
+    "calciteAvatica": "org.apache.calcite:calcite-avatica:1.3.0-incubating",
     "jhyde": "org.pentaho:pentaho-aggdesigner-algorithm:5.1.5-jhyde",
     "curatorFramework": "org.apache.curator:curator-framework:2.10.0",
     "curatorRecipes": "org.apache.curator:curator-recipes:2.10.0",
